basic-python/requirements.txt
requests==2.18.4
click==6.7
PyYAML==3.12
